fix: split encountered skills command
