+++ crates/tracepilot-tauri-bindings/src/lib.rs
@@ -286,7 +286,7 @@ pub fn init() -> tauri::plugin::TauriPlugin<tauri::Wry> {
             commands::mcp::mcp_compute_diff,
             // Skills commands (20)
             commands::skills::skills_list_all,
-            commands::skills::skills_encountered_project,
+            commands::skills::encountered::skills_encountered_project,
             commands::skills::skills_get_skill,
             commands::skills::skills_create,
             commands::skills::skills_update,
